Allow inserting elements in custom order
In order to allow more flexibility for the consumer, add an optional
"position" parameter to addElement(); if not specified, the element will
be added to the last position as before.

An example use case is being able to add an element to the bottom of the
stack instead of the top. Another would be allowing reverse z-order,
like new elements always go below the exiting ones.